+++ src/LamaPon/Editor/EditorLayer.h
@@ -38,6 +38,7 @@
 namespace LamaPon
 {
     class BgmLoopPanel;
+    class EditorGuiRenderer;
     class GameExportDialog;
     class GraphicsDevice;
     class PlayerPrefs;
@@ -1117,7 +1118,7 @@ namespace LamaPon
         void ReapplyRemoteMousePosition();
         inline static bool s_normalModeRestartRequested{};
         bool m_win32Initialized{};
-        bool m_dx11Initialized{};
+        std::unique_ptr<EditorGuiRenderer> m_editorGuiRenderer;
         bool m_projectSettingsDialogRequested{};
         bool m_helpCenterRequested{};
         std::string m_imguiIniPath;